Update README.md
## Key Technologies 🔧

### Using SAS 4GL

SAS 4GL is a high-level programming language developed specifically for users working in the SAS environment. Its purpose is to enable easy data processing and report generation, and to allow users to write programs with minimal need to manage technical details. 4GL differs from traditional programming languages, such as SAS Base, in that it focuses on providing simplicity of use and efficiency, eliminating the need to write complex code. It is mainly used for data manipulation, including aggregation, transformation and summary generation, making it a very useful tool for analyzing large data sets and automating repetitive tasks. It simplifies data processing and report generation by abstracting technical complexities, allowing users to:
- Perform data aggregation, transformation, and summarization
- Automate repetitive data tasks
- Reduce manual coding efforts
- Enhance efficiency in data manipulation


### SAS Macro

SAS Macro is a macro language in the SAS environment that allows automation and parameterization of code. By using macros, it is possible to generate dynamic programs that can be adapted to different data and conditions. Macros in SAS allow the creation of repeatable blocks of code that can be easily used in different parts of the program, saving time and reducing coding errors. The language is particularly useful for processing large data sets and for tasks that require the same operations to be used repeatedly, making it an essential tool in data analysis automation projects. It enables:
- Dynamic program generation based on variable inputs
- Code reuse and modularization
- Reduction of repetitive coding tasks
- Improved efficiency in processing large datasets

### Using SAS Flows

SAS Flow is a visual tool used to design and automate analysis processes in the SAS environment. It allows users to create data analysis processes by creating workflows that define the various steps of analysis, such as importing data, processing it, analyzing it and presenting the results. SAS Flow is particularly useful in the context of projects that require automation of repetitive tasks or data processing. Through its use, processes can be designed graphically, making it easy for even those without advanced programming knowledge to build and manage complex analysis and business processes. It offers:
- A graphical interface for creating and organizing analytical processes
- Seamless integration of data import, transformation, analysis, and visualization
- Enhanced accessibility for non-programmers to implement complex analytical tasks